Exchange idea of translating to native/studying pairs.
It is easier to allow vocabulary to be grouped by the language that the
user is studying, and the language that the user is native in, rather
than grouping the vocabulary by the language from/to which the user is
translating. This means that a single tongue pair can contain, for
example, Chinese words to translate to English and English words to
translate to Chinese.