Correcting question-validity check.
It is best to calculate this in the component that uses validity, given
that all information needed to determine this is available. Previously,
there was a bug which incorrectly assessed a form to add a question as
valid if an answer was typed in, even if no question text was entered.